loginpage/makemytrip_automation/requirements.txt
selenium>=4.18.0
undetected-chromedriver>=3.5.5
pytest>=8.0.0
pytest-html>=4.1.1
PyYAML>=6.0.1
Pillow>=10.2.0
webdriver-manager>=4.0.1
